+++ mpl-domain/src/test/java/com/codeit/sb02mplteam2/domain/playlist/service/BasicPlaylistServiceTest.java
@@ -9,12 +9,11 @@ import static org.mockito.Mockito.when;
 import com.codeit.sb02mplteam2.domain.playlist.dto.PlaylistDto;
 import com.codeit.sb02mplteam2.domain.playlist.dto.request.PlaylistCreateRequest;
 import com.codeit.sb02mplteam2.domain.playlist.dto.request.PlaylistUpdateRequest;
-import com.codeit.sb02mplteam2.domain.playlist.dto.request.SubscribeRequest;
 import com.codeit.sb02mplteam2.domain.playlist.entity.Playlist;
-import com.codeit.sb02mplteam2.domain.playlist.entity.Subscribe;
 import com.codeit.sb02mplteam2.domain.playlist.repository.PlaylistRepository;
-import com.codeit.sb02mplteam2.domain.playlist.repository.SubscribeRepository;
 import com.codeit.sb02mplteam2.domain.social.repository.FollowRepository;
+import com.codeit.sb02mplteam2.domain.subscribe.entity.Subscribe;
+import com.codeit.sb02mplteam2.domain.subscribe.repository.SubscribeRepository;
 import com.codeit.sb02mplteam2.domain.user.entity.User;
 import com.codeit.sb02mplteam2.domain.user.repository.UserRepository;
 import java.util.Optional;
@@ -103,69 +102,69 @@ class BasicPlaylistServiceTest {
     );
   }
 
-  @Test
-  @DisplayName("구독 성공 테스트")
-  @Disabled
-  void subscribe() {
-    // given
-    User newUser = new User();
-    when(userRepository.findById(1L)).thenReturn(Optional.of(newUser));
-    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
-    SubscribeRequest request = new SubscribeRequest(1L);
-    // when
-    PlaylistDto playlistDto = playlistService.subscribe(1L, request);
+//  @Test
+//  @DisplayName("구독 성공 테스트")
+//  @Disabled
+//  void subscribe() {
+//    // given
+//    User newUser = new User();
+//    when(userRepository.findById(1L)).thenReturn(Optional.of(newUser));
+//    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
+//    SubscribeRequest request = new SubscribeRequest(1L);
+//    // when
+//    PlaylistDto playlistDto = playlistService.subscribe(1L, request);
+//
+//    // then
+//    assertAll(
+//        () -> assertEquals(title, playlistDto.title()),
+//        () -> assertEquals(description, playlistDto.description()),
+//        () -> assertEquals(2, playlistDto.subscriberCount())
+//    );
+//  }
 
-    // then
-    assertAll(
-        () -> assertEquals(title, playlistDto.title()),
-        () -> assertEquals(description, playlistDto.description()),
-        () -> assertEquals(2, playlistDto.subscriberCount())
-    );
-  }
+//  @Test
+//  @DisplayName("구독 취소 성공 테스트")
+//  @Disabled
+//  void unSubscribe() {
+//    // given
+//    User newUser = new User();
+//    Subscribe newSubscribe = new Subscribe(newUser, playlist);
+//    playlist.subscribe(newSubscribe);
+//
+//    when(userRepository.findById(1L)).thenReturn(Optional.of(newUser));
+//    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
+//    when(subscribeRepository.findByUserAndPlaylist(newUser, playlist)).thenReturn(
+//        Optional.of(newSubscribe));
+//
+//    SubscribeRequest request = new SubscribeRequest( 1L);
+//    // when
+//    PlaylistDto playlistDto = playlistService.unSubscribe(1L, request);
+//    // then
+//    assertAll(
+//        () -> assertEquals(title, playlistDto.title()),
+//        () -> assertEquals(description, playlistDto.description()),
+//        () -> assertEquals(1, playlistDto.subscriberCount())
+//    );
+//  }
 
-  @Test
-  @DisplayName("구독 취소 성공 테스트")
-  @Disabled
-  void unSubscribe() {
-    // given
-    User newUser = new User();
-    Subscribe newSubscribe = new Subscribe(newUser, playlist);
-    playlist.subscribe(newSubscribe);
-
-    when(userRepository.findById(1L)).thenReturn(Optional.of(newUser));
-    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
-    when(subscribeRepository.findByUserAndPlaylist(newUser, playlist)).thenReturn(
-        Optional.of(newSubscribe));
-
-    SubscribeRequest request = new SubscribeRequest( 1L);
-    // when
-    PlaylistDto playlistDto = playlistService.unSubscribe(1L, request);
-    // then
-    assertAll(
-        () -> assertEquals(title, playlistDto.title()),
-        () -> assertEquals(description, playlistDto.description()),
-        () -> assertEquals(1, playlistDto.subscriberCount())
-    );
-  }
-
-  @Test
-  @DisplayName("자기 자신 구독 취소 실패 테스트")
-  void unSubscribeWithMeMustFailed() {
-    // given
-    when(userRepository.findById(1L)).thenReturn(Optional.of(user));
-    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
-    when(subscribeRepository.findByUserAndPlaylist(user, playlist)).thenReturn(
-        Optional.of(subscribe));
-    SubscribeRequest request = new SubscribeRequest(1L);
-    // when
-    PlaylistDto playlistDto = playlistService.unSubscribe(1L, request);
-    // then
-    assertAll(
-        () -> assertEquals(title, playlistDto.title()),
-        () -> assertEquals(description, playlistDto.description()),
-        () -> assertEquals(1, playlistDto.subscriberCount())
-    );
-  }
+//  @Test
+//  @DisplayName("자기 자신 구독 취소 실패 테스트")
+//  void unSubscribeWithMeMustFailed() {
+//    // given
+//    when(userRepository.findById(1L)).thenReturn(Optional.of(user));
+//    when(playlistRepository.findById(1L)).thenReturn(Optional.of(playlist));
+//    when(subscribeRepository.findByUserAndPlaylist(user, playlist)).thenReturn(
+//        Optional.of(subscribe));
+//    SubscribeRequest request = new SubscribeRequest(1L);
+//    // when
+//    PlaylistDto playlistDto = playlistService.unSubscribe(1L, request);
+//    // then
+//    assertAll(
+//        () -> assertEquals(title, playlistDto.title()),
+//        () -> assertEquals(description, playlistDto.description()),
+//        () -> assertEquals(1, playlistDto.subscriberCount())
+//    );
+//  }
 
   @Test
   @DisplayName("업데이트 성공 테스트")
